fixed instr processing
code causes seg fault

+++ lib/runtime/internal/Exe.h
@@ -34,14 +34,14 @@ struct mi32_t {
     SET_mi32B(mi,i); \
 
 
-#define GET_mi64(A, B) (((int64_t)A << 32) | ((int64_t)B & 0xFFFFFFFF))
+#define GET_mi64(A, B) ((((int64_t)A << 32)) | ((int64_t)B & 0xFFFFFFFF))
 
-#define GET_mi32w(i) ((int8_t)(i >> 24))
-#define GET_mi32x(i) ((int8_t)((i >> 16) & 0xff))
-#define GET_mi32y(i) ((int8_t)((i >> 8) & 0xff ))
-#define GET_mi32z(i) ((int8_t)(i & 0xff))
+#define GET_mi32w(i) ((uint8_t)(i >> 24))
+#define GET_mi32x(i) ((uint8_t)((i >> 16) & 0xff))
+#define GET_mi32y(i) ((uint8_t)((i >> 8) & 0xff ))
+#define GET_mi32z(i) ((uint8_t)(i & 0xff))
 
-#define SET_mi32(w,x,y,z) ((z) | ((int32_t)y << 8) | ((int32_t)x << 16) | ((int32_t)w << 24))
+#define SET_mi32(w,x,y,z) (((uint8_t)z) | ((uint8_t)y << 8) | ((uint8_t)x << 16) | ((uint8_t)w << 24))
 
 int Process_Exe(std::string);
